refactor: split oversized UI files by responsibility
Pure code motion, no behavior change. SettingsView.swift (1202 lines)
splits into the root view plus SettingsPreferenceSections,
SettingsConnectionSections, and SettingsComponents; IslandWindow.swift
(911) moves the system-overview state machine and the outside-click/
auto-collapse handling into IslandWindow+SystemOverview and
IslandWindow+AutoCollapse extensions (following the SessionStore+*
precedent); IslandDashboardViews.swift (843) moves the approval cards
into IslandApprovalViews. Members shared across the new files drop
`private` to internal; moved section structs likewise.

Verified as a pure move: normalized line-multiset comparison of old vs
new files is exactly conserved (2253 = 2253), and the window
automation suite passes (single approval, approval queue,
expand/collapse, idle, stale-event aging), with the relocated
system-overview hide/restore state machine exercised directly via a
distributed notification (hidden -> force restored in the log).
