+++ Sources/VibelslandFree/IslandWindow.swift
@@ -6,30 +6,30 @@ import SwiftUI
 
 final class IslandWindow: NSPanel {
     private var cancellables: Set<AnyCancellable> = []
-    private var outsideClickMonitor: Any?
-    private var outsideClickArmedAt: Date?
-    private var autoCollapseTimer: Timer?
-    private var autoCollapseWatchActive = false
-    private var systemOverviewTriggerMonitor: Any?
-    private var systemOverviewDetectionTimer: Timer?
-    private var systemOverviewDetectionTicks = 0
-    private var systemOverviewTimer: Timer?
-    private var systemOverviewEventMonitor: Any?
-    private var systemOverviewMinimumRestoreAt: Date?
-    private var systemOverviewForceRestoreAt: Date?
-    private var systemOverviewRestoreFrame: NSRect?
-    private var preferredCenterXForNextFrame: CGFloat?
+    var outsideClickMonitor: Any?
+    var outsideClickArmedAt: Date?
+    var autoCollapseTimer: Timer?
+    var autoCollapseWatchActive = false
+    var systemOverviewTriggerMonitor: Any?
+    var systemOverviewDetectionTimer: Timer?
+    var systemOverviewDetectionTicks = 0
+    var systemOverviewTimer: Timer?
+    var systemOverviewEventMonitor: Any?
+    var systemOverviewMinimumRestoreAt: Date?
+    var systemOverviewForceRestoreAt: Date?
+    var systemOverviewRestoreFrame: NSRect?
+    var preferredCenterXForNextFrame: CGFloat?
     private var lastAppliedPosition: IslandPosition?
     private var lastAppliedExpanded: Bool?
     private var lastLayoutSignature: IslandLayoutSignature?
-    private var lastVisibleFrame: NSRect?
+    var lastVisibleFrame: NSRect?
     private var transitionResetTask: Task<Void, Never>?
     private var frameAnimationTimer: Timer?
     private var hasPresented = false
-    private var hiddenForSystemOverview = false
-    private var suppressedForSettings = false
-    private weak var store: SessionStore?
-    private let logger = AppLogger.shared
+    var hiddenForSystemOverview = false
+    var suppressedForSettings = false
+    weak var store: SessionStore?
+    let logger = AppLogger.shared
 
     override var canBecomeKey: Bool { true }
     override var canBecomeMain: Bool { true }
@@ -198,223 +198,7 @@ final class IslandWindow: NSPanel {
         startSystemOverviewTriggerMonitor()
     }
 
-    private func startSystemOverviewDetectionTimer() {
-        guard systemOverviewDetectionTimer == nil else { return }
-        systemOverviewDetectionTicks = 0
-        let timer = Timer(timeInterval: 0.22, repeats: true) { [weak self] _ in
-            DispatchQueue.main.async {
-                guard let self, !self.hiddenForSystemOverview else { return }
-                self.systemOverviewDetectionTicks += 1
-                if self.systemOverviewLikelyVisible() {
-                    self.hideForSystemOverview(minimumDuration: 0.45)
-                    self.stopSystemOverviewDetectionTimer()
-                    return
-                }
-                if self.systemOverviewDetectionTicks >= 12 {
-                    self.stopSystemOverviewDetectionTimer()
-                }
-            }
-        }
-        systemOverviewDetectionTimer = timer
-        RunLoop.main.add(timer, forMode: .common)
-    }
-
-    private func stopSystemOverviewDetectionTimer() {
-        systemOverviewDetectionTimer?.invalidate()
-        systemOverviewDetectionTimer = nil
-        systemOverviewDetectionTicks = 0
-    }
-
-    private func startSystemOverviewTriggerMonitor() {
-        guard systemOverviewTriggerMonitor == nil else { return }
-        systemOverviewTriggerMonitor = NSEvent.addGlobalMonitorForEvents(
-            matching: [.keyDown, .swipe, .gesture]
-        ) { [weak self] event in
-            DispatchQueue.main.async {
-                self?.handlePotentialSystemOverviewEvent(event)
-            }
-        }
-    }
-
-    private func handlePotentialSystemOverviewEvent(_ event: NSEvent) {
-        switch event.type {
-        case .swipe, .gesture:
-            startSystemOverviewDetectionTimer()
-            hideForSystemOverview(minimumDuration: 0.9)
-        case .keyDown:
-            let isMissionControlShortcut =
-                event.keyCode == 99 ||
-                (event.keyCode == 126 && event.modifierFlags.contains(.control))
-            if isMissionControlShortcut {
-                startSystemOverviewDetectionTimer()
-                hideForSystemOverview(minimumDuration: 0.9)
-            }
-        default:
-            break
-        }
-    }
-
-    private func handleActiveApplicationChanged(_ notification: Notification) {
-        let application = notification.userInfo?[NSWorkspace.applicationUserInfoKey] as? NSRunningApplication
-        guard application?.bundleIdentifier == "com.apple.dock" else {
-            restoreAfterSystemOverviewIfNeeded()
-            return
-        }
-        hideForSystemOverview(minimumDuration: 0.7)
-    }
-
-    private func hideForSystemOverview(minimumDuration: TimeInterval) {
-        let now = Date()
-        let minimumRestoreAt = now.addingTimeInterval(minimumDuration)
-        let forceRestoreAt = now.addingTimeInterval(max(minimumDuration + 2.4, 3.0))
-        if let current = systemOverviewMinimumRestoreAt {
-            systemOverviewMinimumRestoreAt = max(current, minimumRestoreAt)
-        } else {
-            systemOverviewMinimumRestoreAt = minimumRestoreAt
-        }
-        if let current = systemOverviewForceRestoreAt {
-            systemOverviewForceRestoreAt = max(current, forceRestoreAt)
-        } else {
-            systemOverviewForceRestoreAt = forceRestoreAt
-        }
-        startSystemOverviewEventMonitor()
-        startSystemOverviewTimer()
-        stopSystemOverviewDetectionTimer()
-        guard !hiddenForSystemOverview else { return }
-        let restoreFrame = lastVisibleFrame ?? frame
-        if restoreFrame.width > 1, restoreFrame.height > 1 {
-            systemOverviewRestoreFrame = restoreFrame
-        }
-        hiddenForSystemOverview = true
-        logger.info("island.system-overview.hidden")
-        alphaValue = 0
-        setFrame(systemOverviewHiddenFrame(), display: false)
-        orderOut(nil)
-        stopOutsideClickMonitor()
-        stopAutoCollapseWatch()
-    }
-
-    private func systemOverviewHiddenFrame() -> NSRect {
-        let screenFrame = NSScreen.main?.frame ?? .zero
-        return NSRect(x: screenFrame.maxX + 200, y: screenFrame.maxY + 200, width: 1, height: 1)
-    }
-
-    private func startSystemOverviewEventMonitor() {
-        guard systemOverviewEventMonitor == nil else { return }
-        systemOverviewEventMonitor = NSEvent.addGlobalMonitorForEvents(
-            matching: [.leftMouseUp, .rightMouseUp, .otherMouseUp, .keyUp, .swipe]
-        ) { [weak self] _ in
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.18) {
-                self?.restoreAfterSystemOverviewIfNeeded()
-            }
-        }
-    }
-
-    private func stopSystemOverviewEventMonitor() {
-        if let systemOverviewEventMonitor {
-            NSEvent.removeMonitor(systemOverviewEventMonitor)
-            self.systemOverviewEventMonitor = nil
-        }
-    }
-
-    private func startSystemOverviewTimer() {
-        systemOverviewTimer?.invalidate()
-        let timer = Timer(timeInterval: 0.18, repeats: true) { [weak self] _ in
-            DispatchQueue.main.async {
-                self?.restoreAfterSystemOverviewIfNeeded()
-            }
-        }
-        systemOverviewTimer = timer
-        RunLoop.main.add(timer, forMode: .common)
-    }
-
-    private func stopSystemOverviewTimer() {
-        systemOverviewTimer?.invalidate()
-        systemOverviewTimer = nil
-    }
-
-    private func restoreAfterSystemOverviewIfNeeded(force: Bool = false) {
-        guard hiddenForSystemOverview else { return }
-        guard !suppressedForSettings else { return }
-        let now = Date()
-        let decision = SystemOverviewRestorePolicy.decision(
-            force: force,
-            now: now,
-            minimumRestoreAt: systemOverviewMinimumRestoreAt,
-            forceRestoreAt: systemOverviewForceRestoreAt,
-            overviewLikelyVisible: systemOverviewLikelyVisible(),
-            frontmostBundleID: NSWorkspace.shared.frontmostApplication?.bundleIdentifier
-        )
-        switch decision {
-        case .wait:
-            return
-        case .forceRestore:
-            logger.info("island.system-overview.restore.forced")
-        case .restore:
-            break
-        }
-        hiddenForSystemOverview = false
-        systemOverviewMinimumRestoreAt = nil
-        systemOverviewForceRestoreAt = nil
-        logger.info("island.system-overview.restored")
-        stopSystemOverviewTimer()
-        stopSystemOverviewEventMonitor()
-        stopSystemOverviewDetectionTimer()
-        guard let store else { return }
-        if let systemOverviewRestoreFrame {
-            preferredCenterXForNextFrame = systemOverviewRestoreFrame.midX
-            setFrame(systemOverviewRestoreFrame, display: false)
-            self.systemOverviewRestoreFrame = nil
-        }
-        alphaValue = 1
-        applyFrame(
-            expanded: store.isExpanded,
-            position: store.configurationStore.config.islandPosition,
-            animated: false
-        )
-        if store.isExpanded {
-            makeKeyAndOrderFront(nil)
-        } else {
-            orderFrontRegardless()
-        }
-        repairFrameIfNeeded()
-    }
-
-    private func systemOverviewLikelyVisible() -> Bool {
-        guard let screen = NSScreen.screens.first(where: { $0.frame.intersects(frame) }) ?? NSScreen.main,
-              let windowList = CGWindowListCopyWindowInfo([.optionOnScreenOnly, .excludeDesktopElements], kCGNullWindowID) as? [[String: Any]] else {
-            return false
-        }
-        let screenFrame = screen.frame
-        return windowList.contains { info in
-            guard isDockWindow(info) else {
-                return false
-            }
-            let name = info[kCGWindowName as String] as? String ?? ""
-            if name.localizedCaseInsensitiveContains("mission") ||
-                name.localizedCaseInsensitiveContains("expose") ||
-                name.localizedCaseInsensitiveContains("spaces") {
-                return true
-            }
-            guard let bounds = info[kCGWindowBounds as String] as? [String: Any],
-                  let width = bounds["Width"] as? CGFloat,
-                  let height = bounds["Height"] as? CGFloat else {
-                return false
-            }
-            return width > screenFrame.width * 0.55 && height > screenFrame.height * 0.45
-        }
-    }
-
-    private func isDockWindow(_ info: [String: Any]) -> Bool {
-        if let ownerPID = info[kCGWindowOwnerPID as String] as? pid_t,
-           NSRunningApplication(processIdentifier: ownerPID)?.bundleIdentifier == "com.apple.dock" {
-            return true
-        }
-        let ownerName = info[kCGWindowOwnerName as String] as? String ?? ""
-        return ownerName == "Dock" || ownerName == "程序坞"
-    }
-
-    private func repairFrameIfNeeded(shouldOrder: Bool = true) {
+    func repairFrameIfNeeded(shouldOrder: Bool = true) {
         guard let store else { return }
         let target = targetFrame(
             expanded: store.isExpanded,
@@ -440,7 +224,7 @@ final class IslandWindow: NSPanel {
         )
     }
 
-    private func applyFrame(
+    func applyFrame(
         expanded: Bool,
         position: IslandPosition,
         animated: Bool,
@@ -704,117 +488,6 @@ final class IslandWindow: NSPanel {
             maximumHeight = hasHealthWarning ? 376 : 342
         }
         return min(max(height, 126), maximumHeight)
-    }
-
-    private func updateOutsideClickMonitor(expanded: Bool) {
-        if hiddenForSystemOverview {
-            stopOutsideClickMonitor()
-            stopAutoCollapseWatch()
-            return
-        }
-        if expanded {
-            startOutsideClickMonitor()
-            if hasPendingApproval {
-                stopAutoCollapseWatch()
-            } else {
-                startAutoCollapseWatch()
-            }
-        } else {
-            stopOutsideClickMonitor()
-            stopAutoCollapseWatch()
-        }
-    }
-
-    private func startOutsideClickMonitor() {
-        guard outsideClickMonitor == nil else { return }
-        outsideClickArmedAt = Date().addingTimeInterval(0.8)
-        outsideClickMonitor = NSEvent.addGlobalMonitorForEvents(
-            matching: [.leftMouseDown, .rightMouseDown, .otherMouseDown]
-        ) { [weak self] _ in
-            DispatchQueue.main.async {
-                self?.collapseIfOutsideClick()
-            }
-        }
-    }
-
-    private func stopOutsideClickMonitor() {
-        outsideClickArmedAt = nil
-        if let outsideClickMonitor {
-            NSEvent.removeMonitor(outsideClickMonitor)
-            self.outsideClickMonitor = nil
-        }
-    }
-
-    private func collapseIfOutsideClick() {
-        guard let store, store.isExpanded else { return }
-        if let outsideClickArmedAt, Date() < outsideClickArmedAt {
-            return
-        }
-        if frame.insetBy(dx: -2, dy: -2).contains(NSEvent.mouseLocation) {
-            return
-        }
-        collapseIfExpanded()
-    }
-
-    private func collapseIfExpanded() {
-        guard let store, store.isExpanded else { return }
-        store.isExpanded = false
-    }
-
-    /// 事件驱动的离开收起：tracking area 的进出事件代替旧的 0.22 秒鼠标轮询。
-    /// 鼠标离开（或启动监视时就在窗外）后启动一次性宽限定时器，重新进入即取消。
-    private func startAutoCollapseWatch() {
-        autoCollapseWatchActive = true
-        if !frame.insetBy(dx: -12, dy: -12).contains(NSEvent.mouseLocation) {
-            armAutoCollapseTimer()
-        }
-    }
-
-    private func stopAutoCollapseWatch() {
-        autoCollapseWatchActive = false
-        autoCollapseTimer?.invalidate()
-        autoCollapseTimer = nil
-    }
-
-    override func mouseEntered(with event: NSEvent) {
-        autoCollapseTimer?.invalidate()
-        autoCollapseTimer = nil
-        super.mouseEntered(with: event)
-    }
-
-    override func mouseExited(with event: NSEvent) {
-        if autoCollapseWatchActive {
-            armAutoCollapseTimer()
-        }
-        super.mouseExited(with: event)
-    }
-
-    private func armAutoCollapseTimer() {
-        guard autoCollapseWatchActive, store?.isExpanded == true, !hasPendingApproval else { return }
-        autoCollapseTimer?.invalidate()
-        let timer = Timer(timeInterval: IslandAutoCollapsePolicy.graceDuration, repeats: false) { [weak self] _ in
-            DispatchQueue.main.async {
-                self?.fireAutoCollapse()
-            }
-        }
-        RunLoop.main.add(timer, forMode: .common)
-        autoCollapseTimer = timer
-    }
-
-    private func fireAutoCollapse() {
-        autoCollapseTimer = nil
-        guard autoCollapseWatchActive, let store, store.isExpanded, !hasPendingApproval else { return }
-        // 窗口动画期间可能漏掉进入事件，收起前再确认鼠标确实在窗外。
-        guard !frame.insetBy(dx: -12, dy: -12).contains(NSEvent.mouseLocation) else { return }
-        collapseIfExpanded()
-    }
-
-    private var hasPendingApproval: Bool {
-        guard let store else { return false }
-        return store.sessions.contains { session in
-            guard let approval = session.approval else { return false }
-            return !approval.isExpired
-        }
     }
 
     private func targetScreen() -> NSScreen? {
